backend/statistical_analysis.py
```python
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.ensemble import RandomForestRegressor
from sklearn.metrics import mean_absolute_error

# Load the data
def load_data():
    global statisticalData
    statisticalData = pd.read_csv('data/statistical_data.csv')

# Preprocess the data
def preprocess_data():
    global X, y
    # Select target
    y = statisticalData.target
    # Select predictors
    X = statisticalData.drop(['target'], axis=1)

# Split the data into training and validation sets
def split_data():
    global train_X, val_X, train_y, val_y
    train_X, val_X, train_y, val_y = train_test_split(X, y, random_state=0)

# Define the model
def define_model():
    global statistical_model
    statistical_model = RandomForestRegressor(random_state=1)

# Fit the model
def fit_model():
    statistical_model.fit(train_X, train_y)

# Make validation predictions and calculate mean absolute error
def validate_model():
    val_predictions = statistical_model.predict(val_X)
    val_mae = mean_absolute_error(val_predictions, val_y)
    print("Validation MAE: {:,.0f}".format(val_mae))

# Conduct statistical analysis
def conductStatisticalAnalysis():
    load_data()
    preprocess_data()
    split_data()
    define_model()
    fit_model()
    validate_model()

if __name__ == "__main__":
    conductStatisticalAnalysis()
```